requirements.txt
dynamiq
python-dotenv>=1.0.1,<2.0.0
